src/main/java/music/controller/AddSongRequest.java
// package
package music.controller;


public class AddSongRequest {
	// variables
	private String title;
	private String artist;
	private String name;

	// no args constructor
	public AddSongRequest() {
		super();
	}

	// getters and setters
	/**
	 * @return title
	 */
	public String getTitle() {
		return title;
	}

	/**
	 * @param title
	 */
	public void setTitle(String title) {
		this.title = title;
	}

	/**
	 * @return artist
	 */
	public String getArtist() {
		return artist;
	}

	/**
	 * @param artist
	 */
	public void setArtist(String artist) {
		this.artist = artist;
	}

	/**
	 * @return name
	 */
	public String getName() {
		return name;
	}

	/**
	 * @param name
	 */
	public void setName(String name) {
		this.name = name;
	}

}
